+++ TShockAPI/GetDataHandlers.cs
@@ -152,6 +152,7 @@ namespace TShockAPI
 					{ PacketTypes.PlayerHurtV2, HandlePlayerDamageV2 },
 					{ PacketTypes.PlayerDeathV2, HandlePlayerKillMeV2 },
 					{ PacketTypes.Emoji, HandleEmoji },
+					{ PacketTypes.SyncTilePicking, HandleSyncTilePicking },
 					{ PacketTypes.SyncRevengeMarker, HandleSyncRevengeMarker },
 					{ PacketTypes.LandGolfBallInCup, HandleLandGolfBallInCup },
 					{ PacketTypes.FishOutNPC, HandleFishOutNPC },
@@ -1870,6 +1871,48 @@ namespace TShockAPI
 		}
 
 		/// <summary>
+		/// For use in a SyncTilePicking event.
+		/// </summary>
+		public class SyncTilePickingEventArgs : GetDataHandledEventArgs
+		{
+			/// <summary>
+			/// The player index in the packet, who sends the tile picking data.
+			/// </summary>
+			public byte PlayerIndex { get; set; }
+			/// <summary>
+			/// The X world position of the tile that is being picked.
+			/// </summary>
+			public short TileX { get; set; }
+			/// <summary>
+			/// The Y world position of the tile that is being picked.
+			/// </summary>
+			public short TileY { get; set; }
+			/// <summary>
+			/// The damage that is being dealt on the tile.
+			/// </summary>
+			public byte TileDamage { get; set; }
+		}
+		/// <summary>
+		/// Called when a player hits and damages a tile.
+		/// </summary>
+		public static HandlerList<SyncTilePickingEventArgs> SyncTilePicking = new HandlerList<SyncTilePickingEventArgs>();
+		private static bool OnSyncTilePicking(TSPlayer player, MemoryStream data, byte playerIndex, short tileX, short tileY, byte tileDamage)
+		{
+			if (SyncTilePicking == null)
+				return false;
+
+			var args = new SyncTilePickingEventArgs
+			{
+				PlayerIndex = playerIndex,
+				TileX = tileX,
+				TileY = tileY,
+				TileDamage = tileDamage
+			};
+			SyncTilePicking.Invoke(null, args);
+			return args.Handled;
+		}
+
+
 		/// For use in an Emoji event.
 		/// </summary>
 		public class EmojiEventArgs : GetDataHandledEventArgs
@@ -3653,7 +3696,20 @@ namespace TShockAPI
 
 			return false;
 		}
-    
+
+		private static bool HandleSyncTilePicking(GetDataHandlerArgs args)
+		{
+			byte playerIndex = args.Data.ReadInt8();
+			short tileX = args.Data.ReadInt16();
+			short tileY = args.Data.ReadInt16();
+			byte damage = args.Data.ReadInt8();
+
+			if (OnSyncTilePicking(args.Player, args.Data, playerIndex, tileX, tileY, damage))
+				return true;
+
+			return false;
+		}
+
 		private static bool HandleSyncRevengeMarker(GetDataHandlerArgs args)
 		{
 			int uniqueID = args.Data.ReadInt32();